// soitic-test-repo/src/dashboard-soitic/backend/src/appointments/appointment.enums.ts
export enum StatusAgendamento {
  CONFIRMADO = 'confirmado',
  PENDENTE = 'pendente',
  CANCELADO = 'cancelado',
  EM_ESPERA = 'em espera',
  A_CAMINHO = 'a caminho',
  ATRASADO = 'atrasado',
}

export enum TipoAgendamento {
  PRIMEIRA_CONSULTA = 'primeira consulta',
  RETORNO = 'retorno',
  EXAME = 'exame',
  URGENCIA = 'urgência',
}